% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package1.R
\name{package1}
\alias{package1}
\title{Labs Package}
\description{
Contains various functions from previous labs for newer labs where using this package is necesssary.
}
\author{
Thuan Tran <thuan.m.tran-1@ou.edu>
}
